% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rra.R
\name{rra}
\alias{rra}
\title{rra}
\usage{
rra(R, G, K, price_err = FALSE, annual_error = 1, weight = FALSE)
}
\arguments{
\item{R}{The t*(1+n) dataframe containing the return data with date.
Make sure that the date column is the first column.}

\item{G}{The t*(1+l) dataframe containing the factor proxies with date.
Make sure that the date column is the first column.}

\item{K}{The number of factors you want to extract.}

\item{price_err}{Bool flag of including price error}

\item{alpha}{The annual pricing error e.g. alpha=1 if 1% per year}
}
\value{
The K extracted factors factors($Phi*G$)
}
\description{
The main implementation of reduced-rank approach
by Dashan Huang et al. (MS, 2022)
}
