% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rra_main.R
\name{rra_main}
\alias{rra_main}
\title{rra_main}
\usage{
rra_main(R, G, K)
}
\arguments{
\item{R}{The t*n matrix containing the return data.}

\item{G}{The t*l matrix containing the factor proxies.}

\item{K}{The number of factors you want to extract.}
}
\value{
The K extracted factors factors($Phi*G$)
}
\description{
The main implementation of reduced-rank approach
by Dashan Huang et al. (MS, 2022)
}
