% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rra_pere.R
\name{rra_pere}
\alias{rra_pere}
\title{rra_pere}
\usage{
rra_pere(R, G, K, alpha)
}
\arguments{
\item{R}{The t*n matrix containing the return data.}

\item{G}{The t*l matrix containing the factor proxies.}

\item{K}{The number of factors you want to extract.}

\item{alpha}{The matrix of alpha*std_error of each basis portfolio.}
}
\value{
The K extracted factors($Phi*G$)
}
\description{
The implementation of reduced-rank approach with pricing error
             restrictions by Dashan Huang et al. (MS, 2022)
}
