% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rra_w.R
\name{rra_w}
\alias{rra_w}
\title{rra_w}
\usage{
rra_w(R, G, K, s2)
}
\arguments{
\item{R}{The t*n matrix containing the return data.}

\item{G}{The t*l matrix containing the factor proxies.}

\item{K}{The number of factors you want to extract.}

\item{S2}{The m*m weight matrix getted from POET.}
}
\value{
The K extracted factors($Phi*G$)
}
\description{
The implementation of reduced-rank approach with a changed weight
by Dashan Huang et al. (MS, 2022)
}
